[include & src] fixed a lot more compile errors concerning the constructors

+++ src/main.cpp
@@ -2,6 +2,6 @@
 
 
 int main(int argc, char** argv){
-    nn<f32> n;
+    nn<f64> n();
     return 0;
 }
